krait driver port
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@34239 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: a6b7c78a15dd545792c83a3aa8dd44433b247277

+++ src/linux/universal/linux-4.14/include/linux/fab_scaling.h
@@ -20,7 +20,7 @@
 
 struct fab_scaling_info {
 	struct clk *clk;
-	unsigned long idle_freq;
+	unsigned int idle_freq;
 };
 
 /**
